# services/account/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the account service for PayGidi.",
        "title": "PayGidi Account Service API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "api.paygidi.site",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/verify/bvn-image": {
            "post": {
                "description": "Verify a user's face image against their BVN record.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify BVN Image",
                "parameters": [
                    {
                        "description": "BVN and Image data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.VerifyBVNImageDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "BVN image verified successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Image mismatch",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/verify/email": {
            "post": {
                "description": "Verify the OTP sent to the user's email address.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email OTP",
                "parameters": [
                    {
                        "description": "Email verification data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.VerifyEmailDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email verified successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid OTP or expired",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "User or OTP not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/verify/nin": {
            "post": {
                "description": "Verify a user's National Identification Number (NIN).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify NIN",
                "parameters": [
                    {
                        "description": "NIN data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/validators.VerifyNINDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "NIN verified successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid NIN",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "502": {
                        "description": "Service error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "validators.VerifyBVNImageDto": {
            "type": "object",
            "required": [
                "base64Image",
                "bvn"
            ],
            "properties": {
                "base64Image": {
                    "type": "string"
                },
                "bvn": {
                    "type": "string"
                }
            }
        },
        "validators.VerifyEmailDto": {
            "type": "object",
            "required": [
                "email",
                "forWhat",
                "otp"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "forWhat": {
                    "type": "string",
                    "enum": [
                        "completeRegister",
                        "login",
                        "resetPassword",
                        "twoFactorAuth"
                    ]
                },
                "otp": {
                    "type": "string",
                    "minLength": 4
                }
            }
        },
        "validators.VerifyNINDto": {
            "type": "object",
            "required": [
                "nin"
            ],
            "properties": {
                "nin": {
                    "type": "string",
                    "minLength": 10
                }
            }
        }
    }
}